fix: corrigindo toast e testando com endpoint certo
