+++ src/Pages/Login/index.jsx
@@ -3,16 +3,37 @@ import { GetStatus } from "../../services/GetStatus";
 import style from "./style.module.css";
 import { useForm } from "react-hook-form";
 import { ToastContainer } from "react-toastify";
+import { toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 
 export const Login = () => {
     const { register, handleSubmit } = useForm();
     const navigate = useNavigate();
+
     const onSubmit = async (data) => {
         try {
-            await GetStatus(data);
-            navigate("/home");
+            const response = await GetStatus(data.apiKey);
+            const notify = (msg) => {
+                if (msg === "sucess") {
+                    toast.success("Login Efetuado !", {
+                        position: toast.POSITION.TOP_RIGHT,
+                    });
+                }
+                if (msg === "error") {
+                    toast.error("Api Key inválida !", {
+                        position: toast.POSITION.TOP_RIGHT,
+                    });
+                }
+            };
+            if (response.errors.length === 0) {
+                notify("sucess");
+            } else {
+                notify("error");
+            }
             if (localStorage.length > 0) {
-                navigate("/home");
+                setTimeout(() => {
+                    navigate("/home");
+                }, 2000);
             }
         } catch (error) {
             console.log(error);
