+++ src/services/GetSeasons/index.js
@@ -1,7 +1,7 @@
-//import { API } from "../API/api";
-import seasons from "../../mocks/seasons.json";
+import { API } from "../API/api";
+//import seasons from "../../mocks/seasons.json";
 
-export const GetSeasons = async () => {
+/* export const GetSeasons = async () => {
     try {
         const data = seasons.response;
         const options = data.map((item) => ({
@@ -12,8 +12,8 @@ export const GetSeasons = async () => {
     } catch (error) {
         console.log("", error);
     }
-};
-/* export const GetSeasons = async () => {
+}; */
+export const GetSeasons = async () => {
     const apiKey = localStorage.getItem("apiKey");
     //fake: demo7870822.mockable.io
     //correct: v3.football.api-sports.io
@@ -34,4 +34,4 @@ export const GetSeasons = async () => {
     } catch (error) {
         console.log("", error);
     }
-}; */
+};
